+++ src/main/java/com/example/examplemod/event/ModEvents.java
@@ -1,6 +1,7 @@
 package com.example.examplemod.event;
 
 import com.example.examplemod.ExampleMod;
+import com.example.examplemod.potion.ModPotions;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -8,8 +9,12 @@ import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.animal.sheep.Sheep;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.alchemy.PotionBrewing;
+import net.minecraft.world.item.alchemy.Potions;
+import net.minecraft.world.level.block.Blocks;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.event.brewing.RegisterBrewingRecipesEvent;
 import net.neoforged.neoforge.event.entity.living.LivingDamageEvent;
 
 @EventBusSubscriber(modid = ExampleMod.MOD_ID)
@@ -24,5 +29,15 @@ public class ModEvents {
                 sheep.addEffect(new MobEffectInstance(MobEffects.POISON, 600, 5));
             }
         }
+    }
+
+    @SubscribeEvent
+    public static void onBrewingRecipeRegister(RegisterBrewingRecipesEvent event) {
+        // Example of how to register a brewing recipe
+        // BrewingRecipeRegistry.addRecipe(new ItemStack(Items.POTION),
+        // Ingredient.of(Items.NETHER_WART), new ItemStack(Items.POTION, 1));
+        PotionBrewing.Builder builder = event.getBuilder();
+
+        builder.addMix(Potions.AWKWARD, Blocks.DIRT.asItem(), ModPotions.STINKY_POTION);
     }
 }